DRTII-1261 Update admin export for easier old vs new forecast comparison
