+++ server/src/main/scala/controllers/application/exports/WithFlightsExport.scala
@@ -118,16 +118,7 @@ trait WithFlightsExport {
     (user, _, _) =>
       (start, end, terminal) =>
         if (user.hasRole(SuperAdmin)) {
-          val predictedPreferredOrder = List(
-            ScenarioSimulationSource,
-            LiveFeedSource,
-            ApiFeedSource,
-            MlFeedSource,
-            ForecastFeedSource,
-            HistoricApiFeedSource,
-            AclFeedSource,
-          )
-          AdminExportImpl(start, end, terminal, ctrl.paxFeedSourceOrder, predictedPreferredOrder)
+          AdminExportImpl(start, end, terminal, ctrl.paxFeedSourceOrder)
         } else if (user.hasRole(ApiView))
           FlightsWithSplitsWithActualApiExportImpl(start, end, terminal, ctrl.paxFeedSourceOrder)
         else FlightsWithSplitsWithoutActualApiExportImpl(start, end, terminal, ctrl.paxFeedSourceOrder)
