+++ server/src/main/scala/services/exports/flights/templates/FlightsWithSplitsExport.scala
@@ -3,7 +3,6 @@ package services.exports.flights.templates
 import drt.shared.CrunchApi.MillisSinceEpoch
 import passengersplits.parsing.VoyageManifestParser.VoyageManifest
 import uk.gov.homeoffice.drt.arrivals.{ApiFlightWithSplits, ArrivalExportHeadings}
-import uk.gov.homeoffice.drt.ports.MlFeedSource
 
 
 trait FlightsWithSplitsExport extends FlightsExport {
@@ -39,10 +38,6 @@ trait FlightsWithSplitsExport extends FlightsExport {
   def pcpPax(fws: ApiFlightWithSplits): String =
     if (fws.apiFlight.Origin.isDomesticOrCta) "-"
     else fws.bestPaxSource(paxFeedSourceOrder).getPcpPax.map(_.toString).getOrElse("0")
-
-  def predictedPcpPax(fws: ApiFlightWithSplits): String =
-    if (fws.apiFlight.Origin.isDomesticOrCta) "-"
-    else fws.apiFlight.PassengerSources.get(MlFeedSource).flatMap(p => p.getPcpPax.map(_.toString)).getOrElse("-")
 
   override val headings: String = ArrivalExportHeadings.arrivalWithSplitsHeadings
